+++ test/kdl-hs-test-decoder.hs
@@ -56,7 +56,7 @@ encodeEntry entry =
 
 encodeValueData :: KDL.ValueData -> Aeson.Value
 encodeValueData = \case
-  KDL.Text s -> val "string" (Text.unpack s)
+  KDL.String s -> val "string" (Text.unpack s)
   KDL.Number x -> val "number" (Scientific.formatScientific Scientific.Fixed Nothing x)
   KDL.Bool x -> val "boolean" (if x then "true" else "false")
   KDL.Inf -> val "number" "inf"